requirements.txt
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
gym==0.26.2
gym-notices==0.0.8
imageio==2.36.1
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.12.12
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
typing_extensions==4.9.0
tzdata==2024.2
